Add cron

+++ backend/article.py
@@ -1,28 +1,48 @@
+import os
 import feedparser
 from config import DATA_DIR
 from newspaper import Article
 
-RSS_FEEDS = [
-    "https://doordash.engineering/feed",
+
+# https://backfeed.app
+# https://blog.feedspot.com/engineering_rss_feeds/
+BACKFILLED_RSS_FEEDS = [
+    "https://backfeed.app/eOL3r1ethXnHeNqBEy/https://doordash.engineering/feed"
 ]
 
 
-def download_feed(url: str):
-    feed = feedparser.parse(url)
-    print(f"Found {len(feed.entries)} entries in {url}")
+def download_feed(feed_url: str):
+    print(feed_url)
+
+    feed = feedparser.parse(feed_url)
+    print(f"\tFound {len(feed.entries)} entries in feed")
+
+    n_new = 0
 
     for entry in feed.entries:
-        download_article(entry["link"])
+        article_url = entry["link"]
+        is_new = download_article(article_url)
+        n_new += 1 if is_new else 0
+
+    print(f"\tDownloaded {n_new} new articles")
 
 
-def download_article(url: str):
-    title, text = load_article(url)
+def download_article(url: str) -> bool:
+    title, content = load_article(url)
+
+    if not title or not content:
+        return False
+
     filepath = f"{DATA_DIR}/{title}.txt"
 
-    with open(filepath, "w") as f:
-        f.write(text)
+    if os.path.exists(filepath):
+        return False
 
-    print(f"✅ {title}")
+    with open(filepath, "w") as f:
+        f.write(content)
+
+    print(f"\t✅ {title}")
+    return True
 
 
 def load_article(url) -> tuple[str, str]:
@@ -35,10 +55,5 @@ def load_article(url) -> tuple[str, str]:
 
 
 if __name__ == "__main__":
-    # https://blog.feedspot.com/engineering_rss_feeds/
-    # https://backfeed.app
-    feed_url = "https://doordash.engineering/feed"
-    download_feed(feed_url)
-
-    # article_url = "https://doordash.engineering/2021/07/14/open-source-search-indexing"
-    # download_article(article_url)
+    for feed_url in BACKFILLED_RSS_FEEDS:
+        download_feed(feed_url)
